pyproject.toml
[project]
name = "dora-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.21.0",
    "gradio>=5.49.1",
    "groq>=0.33.0",
    "gtts>=2.5.4",
    "langchain-google-genai>=3.0.1",
    "langgraph>=1.0.2",
    "opencv-python>=4.12.0.88",
    "playsound==1.2.2",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "speechrecognition>=3.14.3",
    "wikipedia>=1.4.0",
]
